#23 Refactor(Authentication): 인증 앱 생성
기존의 'account' 앱에 포함되어 있었던 인증과정을 새로운 앱에 이동하였음.

authentication 앱 디렉토리 아래에 oauth 디렉토리와 local_auth 디렉토리를 생성하였음.

해당 디렉토리 아래에 views.py 파일과 urls.py 파일을 생성하여 해당되는 함수와 API 엔드포인트를 할당하여 코드의 가독성과 유지보수를 용이하게 하였음.
