+++ backend/transcendence/urls.py
@@ -21,5 +21,6 @@ from django.urls import path, include
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("account/", include("account.urls")),
+    path("authentication/", include("authentication.urls")),
     path("follow/", include("follow.urls")),
 ]
